build: include SS data into consolidated
Ref eng-476
