+++ packages/core/src/external/surescripts/command/bundle/get-bundle.ts
@@ -1,27 +1,39 @@
 import { Bundle, BundleEntry } from "@medplum/fhirtypes";
+import { Config } from "../../../../util/config";
+import { S3Utils } from "../../../aws/s3";
 import { buildCollectionBundle } from "../../../fhir/bundle/bundle";
+import { buildLatestConversionBundleFileName } from "../../file/file-names";
 
 /**
- * TODO ENG-476 Implement this
+ * Returns the bundle with Surescripts data for a given patient.
+ *
+ * @param cxId - The ID of the care experience.
+ * @param patientId - The ID of the patient.
+ * @returns The bundle with Surescripts data.
  */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-export async function getPharmacyBundle({
+export async function getBundle({
   cxId,
   patientId,
 }: {
   cxId: string;
   patientId: string;
 }): Promise<Bundle> {
-  return buildCollectionBundle([]);
+  const s3Utils = new S3Utils(Config.getAWSRegion());
+  const bucketName = Config.getPharmacyConversionBucketName();
+  const fileName = buildLatestConversionBundleFileName(cxId, patientId);
+  const fileExists = await s3Utils.fileExists(bucketName, fileName);
+  if (!fileExists) return buildCollectionBundle([]);
+  const fileContents = await s3Utils.getFileContentsAsString(bucketName, fileName);
+  return JSON.parse(fileContents);
 }
 
-export async function getPharmacyResources({
+export async function getBundleResources({
   cxId,
   patientId,
 }: {
   cxId: string;
   patientId: string;
 }): Promise<BundleEntry[]> {
-  const bundle = await getPharmacyBundle({ cxId, patientId });
+  const bundle = await getBundle({ cxId, patientId });
   return bundle.entry ?? [];
 }
